+++ src/soloud/src/backend/miniaudio/soloud_miniaudio.cpp
@@ -63,6 +63,7 @@ namespace SoLoud
 {
     ma_device gDevice;
     SoLoud::Soloud *soloud;
+    ma_context context;
 
     // Added by Marco Bavagnoli
     void on_notification(const ma_device_notification* pNotification)
@@ -96,6 +97,12 @@ namespace SoLoud
 
             case ma_device_notification_type_interruption_ended:
             {
+#if defined(MA_HAS_COREAUDIO)
+                // On macOS and iOS when the the interruption begins
+                // the device is automatically stopped (not uninited with ma_device_uninit).
+                // So we need to start it again when the interruption ends.
+                miniaudio_ensureDeviceStarted_impl();
+#endif
                 soloud->_stateChangedCallback(4);
             } break;
 
@@ -116,30 +123,57 @@ namespace SoLoud
 
     static void soloud_miniaudio_deinit(SoLoud::Soloud *aSoloud)
     {
+        ma_device_stop(&gDevice);
         ma_device_uninit(&gDevice);
+#if defined(MA_HAS_COREAUDIO)
+        ma_context_uninit(&context);
+#endif
     }
 
     result miniaudio_init(SoLoud::Soloud *aSoloud, unsigned int aFlags, unsigned int aSamplerate, unsigned int aBuffer, unsigned int aChannels, void *pPlaybackInfos_id)
     {
         soloud = aSoloud;
-        ma_device_config config = ma_device_config_init(ma_device_type_playback);
+        ma_device_config deviceConfig = ma_device_config_init(ma_device_type_playback);
         if (pPlaybackInfos_id != NULL)
         {
-            config.playback.pDeviceID = (ma_device_id*)pPlaybackInfos_id;
+            deviceConfig.playback.pDeviceID = (ma_device_id*)pPlaybackInfos_id;
         }
-        config.periodSizeInFrames = aBuffer;
-        config.playback.format    = ma_format_f32;
-        config.playback.channels  = aChannels;
-        config.sampleRate         = aSamplerate;
-        config.dataCallback       = soloud_miniaudio_audiomixer;
-        config.pUserData          = (void *)aSoloud;
-        if (aSoloud->_stateChangedCallback != nullptr)
-            config.notificationCallback = on_notification;
+        deviceConfig.periodSizeInFrames = aBuffer;
+        deviceConfig.playback.format    = ma_format_f32;
+        deviceConfig.playback.channels  = aChannels;
+        deviceConfig.sampleRate         = aSamplerate;
+        deviceConfig.dataCallback       = soloud_miniaudio_audiomixer;
+        deviceConfig.pUserData          = (void *)aSoloud;
 
-        if (ma_device_init(NULL, &config, &gDevice) != MA_SUCCESS)
+        // deviceConfig.aaudio.usage       = ma_aaudio_usage_default;
+        // deviceConfig.aaudio.contentType = ma_aaudio_content_type_default;
+        // deviceConfig.aaudio.inputPreset = ma_aaudio_input_preset_default;
+        if (aSoloud->_stateChangedCallback != nullptr)
+            deviceConfig.notificationCallback = on_notification;
+
+#if defined(MA_HAS_COREAUDIO)
+        // Disable CoreAudio context
+        ma_context_config contextConfig = ma_context_config_init();
+        contextConfig.coreaudio.sessionCategory = ma_ios_session_category_none;
+        contextConfig.coreaudio.noAudioSessionActivate = true;
+        contextConfig.coreaudio.noAudioSessionDeactivate = true;
+
+        ma_result result = ma_context_init(NULL, 0, &contextConfig, &context);
+        if (result != MA_SUCCESS) {
+            return UNKNOWN_ERROR;
+        }
+        if (ma_device_init(&context, &deviceConfig, &gDevice) != MA_SUCCESS)
+        {
+            ma_context_uninit(&context);
+            return UNKNOWN_ERROR;
+        }
+#else
+        if (ma_device_init(NULL, &deviceConfig, &gDevice) != MA_SUCCESS)
         {
             return UNKNOWN_ERROR;
         }
+#endif
+
 
         aSoloud->postinit_internal(gDevice.sampleRate, gDevice.playback.internalPeriodSizeInFrames, aFlags, gDevice.playback.channels);
 
@@ -156,19 +190,40 @@ namespace SoLoud
             return UNKNOWN_ERROR;
 
         ma_device_uninit(&gDevice);
-        ma_device_config config = ma_device_config_init(ma_device_type_playback);
-        config.playback.pDeviceID = (ma_device_id *)pPlaybackInfos_id;
-        config.periodSizeInFrames = soloud->mBufferSize;
-        config.playback.format    = ma_format_f32;
-        config.playback.channels  = soloud->mChannels;
-        config.sampleRate         = soloud->mSamplerate;
-        config.dataCallback       = soloud_miniaudio_audiomixer;
-        config.pUserData          = (void *)soloud;
-        if (ma_device_init(NULL, &config, &gDevice) != MA_SUCCESS)
+        ma_device_config deviceConfig = ma_device_config_init(ma_device_type_playback);
+        deviceConfig.playback.pDeviceID = (ma_device_id *)pPlaybackInfos_id;
+        deviceConfig.periodSizeInFrames = soloud->mBufferSize;
+        deviceConfig.playback.format    = ma_format_f32;
+        deviceConfig.playback.channels  = soloud->mChannels;
+        deviceConfig.sampleRate         = soloud->mSamplerate;
+        deviceConfig.dataCallback       = soloud_miniaudio_audiomixer;
+        deviceConfig.pUserData          = (void *)soloud;
+        if (ma_device_init(NULL, &deviceConfig, &gDevice) != MA_SUCCESS)
         {
             return UNKNOWN_ERROR;
         }
         ma_device_start(&gDevice);
+        return 0;
+    }
+
+    // Added to ensure miniaudio device is started when needed, ie by an interruption.
+    // On macOS and iOS when an interruption begins (ie anothe app needs the audio context),
+    // the device is automatically stopped (not uninited with ma_device_uninit).
+    // So we need to check if the device is stopped and start it again.
+    result miniaudio_ensureDeviceStarted_impl()
+    {
+        if (soloud == nullptr)
+            return UNKNOWN_ERROR;
+
+        // Check if device is stopped and start it if needed
+        if (ma_device_get_state(&gDevice) == ma_device_state_stopped)
+        {
+            ma_result result = ma_device_start(&gDevice);
+            if (result != MA_SUCCESS)
+            {
+                return UNKNOWN_ERROR;
+            }
+        }
         return 0;
     }
 };