+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 #### 3.2.0 (XX Xxx 2025)
+- fix #104, #245, #249. Itis now possible to use a 3rd party plugin like `audio_session` to manage audio context.
+- new audio context example in `example/lib/audio_context/audio_context.dart`.
 - fix: GetPosition for buffer streams and Web hot reload/restard #258 and #259
 
 #### 3.1.12 (21 Jun 2025)